getredash/redash#2629 Refactor Chart visualization, add option for handling NULL values (keep/convert to 0.0)
